Create InputFile.py
